Files registering as full updates. Actual changes are mostly limited to modify.js. Added systems to create SQL for series transactions insert & find events of. Still working on update.

+++ source/scripts/transactions/manager.js
@@ -62,18 +62,52 @@ enyo.kind({
 	 */
 	createTransactionFollower: function( data, type, options, results ) {
 
-		var autoTransfer = data['autoTransfer'];
-		var autoTransferLink = data['autoTransferLink'];
-
 		//Set IDs
 		data['itemId'] = parseInt( results[0]['maxItemId'] ) + 1;
 		data['maxRepeatId'] = parseInt( results[0]['maxRepeatId'] ) + 1;
+
+		enyo.application.gts_db.queries(
+				this.generateInsertTransactionSQL( data, type ),
+				{
+					"onSuccess": function() {
+
+						if( enyo.isFunction( options['onSuccess'] ) ) {
+
+							options['onSuccess']( true );
+						}
+					},
+					"onError": function() {
+
+						if( enyo.isFunction( options['onError'] ) ) {
+
+							options['onError']( false );
+						}
+					}
+				}
+			);
+	},
+
+	/**
+	 * @private
+	 * Creates SQL for inserting a new transaction (or set depending on type)
+	 *
+	 * @param {object}	dataIn	Object to be readied; Source object is not modified by function
+	 * @param {string}	type	Type of transaction; used to determine setting of amount;
+	 */
+	generateInsertTransactionSQL: function( dataIn, type ) {
+
+		//Copy object to break pass by reference link
+		var data = enyo.clone( dataIn );
+
+		//Save data about auto transfer system
+		var autoTransfer = data['autoTransfer'];
+		var autoTransferLink = data['autoTransferLink'];
 
 		//Prepare data
 		var sql = this._prepareData( data, type );
 
 		//Handle repeating system
-		sql = sql.concat( this._handleRepeatSystem( data, ( ( autoTransfer > 0 && autoTransferLink >= 0 ) ? autoTransferLink : -1 ) ) );
+		sql = sql.concat( this._handleRepeatSystem( data, autoTransfer, autoTransferLink ) );
 
 		//Handle split transactions
 		sql = sql.concat( this.handleCategoryData( data ) );
@@ -97,25 +131,7 @@ enyo.kind({
 			sql = sql.concat( this.createAutoTransfer( data, autoTransfer, autoTransferLink ) );
 		}
 
-		enyo.application.gts_db.queries(
-				sql,
-				{
-					"onSuccess": function() {
-
-						if( enyo.isFunction( options['onSuccess'] ) ) {
-
-							options['onSuccess']( true );
-						}
-					},
-					"onError": function() {
-
-						if( enyo.isFunction( options['onError'] ) ) {
-
-							options['onError']( false );
-						}
-					}
-				}
-			);
+		return sql;
 	},
 
 	/**
@@ -294,43 +310,6 @@ enyo.kind({
 	},
 
 	/**
-	 * @public
-	 * Checks for repeating transactions
-	 *
-	 * @param int	[acctId]	specific account to update; if not set will update all accounts
-	 * @param {object}	[options]	Callback functions
-	 * @param {function}	[options.onSuccess]
-	 * @param {function}	[options.onError]
-	 */
-	updateSeriesTransactions: function( acctId, options ) {
-
-		//this.seriesDayLimit
-		//this.seriesCountLimit
-/*
-get all series WHERE
-	lastOccurrence < this.seriesDayLimit in the future AND
-	(
-		lastUpdated < current day at 00:00:00 OR
-		lastUpdate IS NULL OR
-		lastUpdate = ''
-	) AND (
-		(
-			endingCondition == 'date' AND
-			endDate < this.seriesDayLimit in the future
-		) OR (
-			endingCondition == 'occurences' AND
-			currCout < endCount
-		)
-	)
-
-///////
-Get max trsn id if results.length > 0
-///////
-loop through series data and create insert statements
-*/
-	},
-
-	/**
 	 * @private
 	 * Modifies the data object to be ready for database insertion. Calls split transaction handler if needed.
 	 *
@@ -346,16 +325,16 @@ loop through series data and create insert statements
 
 		data['amount'] = ( Object.isNumber( data['amount'] ) ? 0 : Number( data['amount'] ).toFixed( 2 ).valueOf() );
 
-		data['date'] =  Date.parse( data['date'] );
+		data['date'] = Date.parse( data['date'] );
 
 		if( type == 'transfer' ) {
 
 			if( data['amount_old'] !== "NOT_A_VALUE" && data['amount_old'] < 0 ) {
-				//Money transferred from here
+				//Existing, money transferred from here
 
 				data['amount'] = -Math.abs( data['amount'] );
 			} else if( data['amount_old'] !== "NOT_A_VALUE" && data['amount_old'] >= 0 ) {
-				//Money transferred to here
+				//Existing, money transferred to here
 
 				data['amount'] = Math.abs( data['amount'] );
 			} else {
@@ -483,29 +462,30 @@ loop through series data and create insert statements
 	 * Handle SQL for repeat events
 	 *
 	 * @param {object}	data	Object to be readied; Since object, pass by reference situation;
-	 * @param [int]	autoTransfer	account to auto transfer to (-1 means no transfer)
+	 * @param [bool]	autoTransfer	Type of autotransfer (0 if none)
+	 * @param [bool]	autoTransferLink	Account to transfer to
 	 *
 	 * @returns {object[]}	SQL functions to be run. Returns empty array if nothing to do.
 	 */
-	_handleRepeatSystem: function( data, autoTransfer ) {
+	_handleRepeatSystem: function( data, autoTransfer, autoTransferLink ) {
 
 		var sqlArray = [];
 
-		if( data['repeatUnlinked'] != 1 ) {
-			//If transaction is not 'unlinked' from repeating series (or not set yet)
+		if( data['rObj'] != false ) {
+			//Not creating from a recursion repeat call
 
-			if( !data['repeatId'] || data['repeatId'] < 0 ) {
+			if( ( !data['repeatId'] || data['repeatId'] < 0 ) && data['rObj']['pattern'] != "none" ) {
 				//New repeating transactions (check/handle)
 
 				data['repeatId'] = data['maxRepeatId'];
 
-				var repeatInsert =  {
+				var repeatInsert = {
 						"repeatId": data['repeatId'],
 
 						//Repeat settings
 						"frequency": data['rObj']['pattern'],//Daily, Weekly, Monthly, Yearly
 						"itemSpan": data['rObj']['frequency'],//Time between events (every 2 months)
-						"daysOfWeek": ( ( data['rObj']['pattern'] == "weekly" ) ? enyo.json.stringify( data['rObj']['dow'] ) : "" ),//Array for day index else blank
+						"daysOfWeek": enyo.json.stringify( ( data['rObj']['pattern'] == "weekly" ) ? data['rObj']['dow'] : "" ),//Array for day index else blank
 
 						"endingCondition": data['rObj']['endCondition'],
 						"endDate": ( ( data['rObj']['endCondition'] == "date" ) ? data['rObj']['endDate'] : "" ),
@@ -513,7 +493,7 @@ loop through series data and create insert statements
 
 						//Initial occurrence is latest and only
 						"lastOccurrence": data['date'],
-						"currCout": 1,
+						"currCount": 1,
 
 						//Original data
 						"origDate": data['rObj']['startDate'],
@@ -523,41 +503,55 @@ loop through series data and create insert statements
 						"rep_category": enyo.json.stringify( data['category'] ),
 						"rep_acctId": data['account'],
 						"rep_linkedAcctId": data['linkedAccount'],
-						"rep_autoTrsnLink": autoTransfer,
+						"rep_autoTrsnLink": ( ( autoTransfer > 0 && autoTransferLink >= 0 ) ? 1 : 0 ),
 
 						//Sync system information
-						"last_sync": ""
+						"last_sync": "",
+
+						//Temp data
+						"maxItemId": ( ( Object.validNumber( data['linkedAccount'] ) && data['linkedAccount'] >= 0 ) ? data['itemId'] + 2 : data['itemId'] + 1 ),
+						"autoTransfer": autoTransfer,
+						"autoTransferLink": autoTransferLink
 					};
 
-				//TODO calculate future repeats (to limit) && update repeatInsert object
+				sqlArray = sqlArray.concat( this.generateSeriesSQL( [ enyo.clone( repeatInsert ) ] ) );
 
-				sqlArray.push( enyo.application.gts_db.getInsert( "repeats", repeatInsert ) );
-			} else if( data['rObj']['pattern'] == "none" ) {
-				//Delete repeating entry
+				//Delete temp data from object
+				delete repeatInsert['maxItemId'];
+				delete repeatInsert['autoTransfer'];
+				delete repeatInsert['autoTransferLink'];
 
-				sqlArray.push( enyo.application.gts_db.getDelete( "repeats", { "repeatId": data['repeatId'] } ) );
+				sqlArray.unshift( enyo.application.gts_db.getInsert( "repeats", repeatInsert ) );
+			} else if( data['repeatUnlinked'] != 1 ) {
+				//If transaction is not 'unlinked' from repeating series
 
-				//Delete future (after curr date) transactions with repeat id except for this one
-				var endOfDay = new Date();
-				endOfDay.setHours( 23, 59, 59, 999 );
+				if( data['rObj']['pattern'] == "none" ) {
+					//Delete repeating entry
 
-				sqlArray.push(
-						new GTS.databaseQuery(
-								{
-									'sql': "DELETE FROM transactions WHERE itemId != ? AND repeatId = ? AND date > ?" ,
-									'values': [ data['itemId'], data['repeatId'], endOfDay.getTime() ]
-								}
-							)
-					);
+					sqlArray.push( enyo.application.gts_db.getDelete( "repeats", { "repeatId": data['repeatId'] } ) );
 
-				//Set repeatId to null
-				data['repeatId'] = null;
-			} else {
-				//Do existing repeating transactions
+					//Delete future (after curr date) transactions with repeat id except for this one
+					var endOfDay = new Date();
+					endOfDay.setHours( 23, 59, 59, 999 );
 
-				//TODO
+					sqlArray.push(
+							new GTS.databaseQuery(
+									{
+										'sql': "DELETE FROM transactions WHERE itemId != ? AND repeatId = ? AND date > ?" ,
+										'values': [ data['itemId'], data['repeatId'], Date.parse( endOfDay ) ]
+									}
+								)
+						);
 
-				//remember to set lastUpdated to null
+					//Set repeatId to null
+					data['repeatId'] = null;
+				} else {
+					//Do existing repeating transactions
+
+					//TODO
+
+					//remember to set lastUpdated to null
+				}
 			}
 		}
 
@@ -572,6 +566,250 @@ loop through series data and create insert statements
 		/////////////////////////////////
 
 		return sqlArray;
+	},
+
+	/**
+	 * @public
+	 * Checks for repeating transactions and creates new transactions if required
+	 *
+	 * @param int	[acctId]	specific account to update; if not set will update all accounts
+	 * @param {object}	[options]	Callback functions
+	 * @param {function}	[options.onSuccess]
+	 * @param {function}	[options.onError]
+	 */
+	updateSeriesTransactions: function( acctId, options ) {
+
+		acctId = acctId || -1;
+
+		var now = new Date();
+		var dayStart = new Date( now.getFullYear(), now.getMonth(), now.getDate(), 0, 0, 0, 0 );
+		var dayLimit = new Date( now.getFullYear(), now.getMonth(), ( now.getDate() + this.seriesDayLimit ), 23, 59, 59, 999 );
+
+		this.generateSeriesSQL( REPEAT_DATA_ARRAY );
+
+		enyo.application.gts_db.query(
+				new GTS.databaseQuery(
+					{
+						'sql': "SELECT * " +
+							"FROM repeats " +
+							"WHERE " +
+								//Account ID
+								"( " +
+									"rep_acctId = ? " +
+									( acctId >= 0 ? "OR 1 = 1 " : "" ) +//Ignore account id if not set
+								") " +
+
+								//Under date limit
+								"lastOccurrence < ? " +
+
+								//Under count limit
+								" AND currCount < ? " +
+
+								//Ending condition checks
+								"AND ( " +
+										"( " +
+											"endingCondition = 'none' " +
+										") OR ( " +
+											"endingCondition = 'date' " +
+											"AND endDate != '' " +
+											"AND endDate > lastOccurrence " +
+										") OR ( " +
+											"endingCondition = 'occurences' " +
+											"AND endCount != '' " +
+											"AND endCount > currCount " +
+										") " +
+									") " +
+
+								//Last data update check
+								"AND ( " +
+									"lastUpdated < ? OR " +
+									"lastUpdate IS NULL OR " +
+									"lastUpdate = '' " +
+								")",
+						'values': [
+							acctId,
+							Date.parse( dayLimit ),
+							this.seriesCountLimit,
+							Date.parse( dayStart )
+						]
+					}
+				),
+				{
+					"onSuccess": function( results ) {
+
+						enyo.application.gts_db.queries( this.generateSeriesSQL( results ), options );
+					},
+					"onError": function() {
+
+						if( enyo.isFunction( options['onError'] ) ) {
+
+							options['onError']();
+						}
+					}
+				}
+			);
+	},
+
+	/**
+	 * @private
+	 * Generates SQL for saving series events
+	 *
+	 * @param {object}[]	repeatArray	Array of objects with series transaction information
+	 *
+	 * @returns {object[]}	SQL functions to be run. Returns empty array if nothing to do.
+	 */
+	generateSeriesSQL: function( repeatArray ) {
+
+		var sql = [];
+
+		if( repeatArray.length > 0 ) {
+
+			//Set max id to build off of; already 1 greater than max id
+			var maxItemId = repeatArray[0]['maxItemId'];
+
+			//Current datetime for calculations
+			var now = new Date();
+
+			//Cycle variables for content
+			var trsnData;
+			var serDate;
+			var serCount;
+			var type;
+
+			//Cycle and create SQL
+			for( var i = 0; i < repeatArray.length; i++ ) {
+
+				serDate = new Date( repeatArray[i]['lastOccurrence'] );
+				serCount = repeatArray[i]['currCount'];
+
+				//Reduce calls to parse function
+				repeatArray[i]['rep_category'] = enyo.json.parse( repeatArray[i]['rep_category'] );
+				repeatArray[i]['daysOfWeek'] = enyo.json.parse( repeatArray[i]['daysOfWeek'] );
+
+				while(
+						//Under date limit
+						Math.floor( ( serDate - new Date( repeatArray[i]['lastOccurrence'] ) ) / ( 1000 * 60 * 60 * 24 ) ) < this.seriesDayLimit
+
+						//Under count limit
+						&& ( serCount - repeatArray[i]['currCount'] ) < this.seriesCountLimit
+
+						//Ending condition checks
+						&& (
+							(
+								repeatArray[i]['endingCondition'] == "none"
+							) || (
+								repeatArray[i]['endingCondition'] == "date"
+								&& repeatArray[i]['endDate']
+								&& Date.parse( serDate ) <= repeatArray[i]['endDate']
+							) || (
+								repeatArray[i]['endingCondition'] == "occurences"
+								&& repeatArray[i]['endCount']
+								&& serCount <= repeatArray[i]['endCount']
+							)
+						)
+					) {
+
+					//Move date to next occurrence
+					if( repeatArray[i]['frequency'] == "daily" ) {
+
+						serDate.setDate( serDate.getDate() + 1 * repeatArray[i]['itemSpan'] );
+					} else if( repeatArray[i]['frequency'] == "weekly" ) {
+
+						//repeatArray[i]['daysOfWeek'] -- array of days repeat occurs on
+
+						var day = repeatArray[i]['daysOfWeek'].indexOf( serDate.getDay() );
+
+						if( day < 0 || ( day + 1 ) >= repeatArray[i]['daysOfWeek'].length ) {
+
+							while( serDate.getDay() > 0 ) {
+								//Increment date until next week (0)
+
+								serDate.setDate( serDate.getDate() + 1 );
+							}
+
+							while( serDate.getDay() < 7 && serDate.getDay() != repeatArray[i]['daysOfWeek'][0] ) {
+								//Increment date until first event of next week
+
+								serDate.setDate( serDate.getDate() + 1 );
+							}
+
+							//Increment date proper number of week spans
+							serDate.setDate( serDate.getDate() + ( 7 * ( repeatArray[i]['itemSpan'] - 1 ) ) );
+						} else {
+
+							//Now + ( number of days between now and next event day )
+							serDate.setDate( serDate.getDate() + ( repeatArray[i]['daysOfWeek'][day + 1] - repeatArray[i]['daysOfWeek'][day] ) );
+							day++
+						}
+					} else if( repeatArray[i]['frequency'] == "monthly" ) {
+
+						serDate.setMonth( serDate.getMonth() + 1 * repeatArray[i]['itemSpan'] );
+					} else if( repeatArray[i]['frequency'] == "yearly" ) {
+
+						serDate.setYear( serDate.getFullYear() + 1 * repeatArray[i]['itemSpan'] );
+					}
+
+					trsnData = {
+						"itemId": maxItemId,
+						"account": repeatArray[i]['rep_acctId'],
+						"repeatId": repeatArray[i]['repeatId'],
+
+						"desc": repeatArray[i]['rep_desc'],
+						"note": repeatArray[i]['rep_note'],
+						"date": Date.parse( serDate ),
+
+						"amount": repeatArray[i]['rep_amount'],
+						"amount_old": "NOT_A_VALUE",
+
+						"linkedRecord": -1,//set elsewhere
+						"linkedAccount": repeatArray[i]['rep_linkedAcctId'],
+
+						"cleared": false,
+						"checkNum": "",
+
+						"category": repeatArray[i]['rep_category'],
+						"category2": "",
+
+						"rObj": false,//do not calculate repeats for this event
+
+						"autoTransfer": ( repeatArray[i]['rep_autoTrsnLink'] == 1 ? repeatArray[i]['autoTransfer'] : 0 ),
+						"autoTransferLink": ( repeatArray[i]['rep_autoTrsnLink'] == 1 ? repeatArray[i]['autoTransferLink'] : -1 )
+					};
+
+					if( Object.validNumber( repeatArray[i]['rep_linkedAcctId'] ) && repeatArray[i]['rep_linkedAcctId'] >= 0 ) {
+
+						type = "transfer";
+					} else if( repeatArray[i]['rep_amount'] < 0 ) {
+
+						type = "expense";
+					} else {
+
+						type = "income";
+					}
+
+					sql = sql.concat( this.generateInsertTransactionSQL( trsnData, type ) );
+
+					serCount++;
+					maxItemId += ( ( Object.validNumber( repeatArray[i]['linkedAccount'] ) && repeatArray[i]['linkedAccount'] >= 0 ) ? 2 : 1 );
+				}
+
+				//update repeat item
+				sql.push(
+						enyo.application.gts_db.getUpdate(
+								"repeats",
+								{
+									"lastOccurrence": Date.parse( serDate ),
+									"currCount": serCount,
+									"lastUpdated": Date.parse( new Date() )
+								}, {
+									"repeatId": repeatArray[i]['repeatId']
+								}
+							)
+					);
+			}
+		}
+
+		return sql;
 	},
 
 	/**
